merge top enrichments and multiple enrichments functionality
The recommend enrichment function can now take as input a number of
enrichments requested. If provided, the tool will return a list of the
top n most appropriate enrichments that are not simultaneous. If not
provided, the tool will return a score for all enrichments batched in
groups of simulaneous enrichments.