fix recommendations of enrichments to use data from IEVCT
Since the data that the IEVCT has changes, we changed the way we figure
out if enrichments are on the same annotation, in order to make the
recomrecommendation on multiple enrichments. We also changes the
information that is passed and displayed in the GUI to much what we have
in our database.